requirements.txt
streamlit
pandas
cloudinary
pymysql
tqdm
stqdm
pillow
